+++ WEB/Src/DependencyCollector/Shared/HttpCoreDiagnosticSourceListener.cs
@@ -47,7 +47,6 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
 
     internal class HttpCoreDiagnosticSourceListener : IObserver<KeyValuePair<string, object>>, IDisposable
     {
-        private const string DependencyErrorPropertyKey = "Error";
         private const string HttpOutEventName = "System.Net.Http.HttpRequestOut";
         private const string HttpOutStartEventName = "System.Net.Http.HttpRequestOut.Start";
         private const string HttpOutStopEventName = "System.Net.Http.HttpRequestOut.Stop";
@@ -341,6 +340,12 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
                 return;
             }
 
+            if (request.Headers.Contains(W3C.W3CConstants.TraceParentHeader) && Activity.DefaultIdFormat == ActivityIdFormat.W3C)
+            {
+                DependencyCollectorEventSource.Log.HttpRequestAlreadyInstrumented(currentActivity.Id);
+                return;
+            }
+
             DependencyCollectorEventSource.Log.HttpCoreDiagnosticSourceListenerStart(currentActivity.Id);
 
             this.InjectRequestHeaders(request, this.configuration.InstrumentationKey);
@@ -364,6 +369,14 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
             if (currentActivity == null)
             {
                 DependencyCollectorEventSource.Log.CurrentActivityIsNull(HttpOutStopEventName);
+                return;
+            }
+
+            if (Activity.DefaultIdFormat == ActivityIdFormat.W3C &&
+                request.Headers.TryGetValues(W3C.W3CConstants.TraceParentHeader, out var parents) && 
+                parents.FirstOrDefault() != currentActivity.Id)
+            {
+                DependencyCollectorEventSource.Log.HttpRequestAlreadyInstrumented();
                 return;
             }
 
@@ -432,7 +445,7 @@ namespace Microsoft.ApplicationInsights.DependencyCollector.Implementation
             {
                 if (this.pendingExceptions.TryRemove(currentActivity.Id, out Exception exception))
                 {
-                    telemetry.Properties[DependencyErrorPropertyKey] = exception.GetBaseException().Message;
+                    telemetry.Properties[RemoteDependencyConstants.DependencyErrorPropertyKey] = exception.GetBaseException().Message;
                 }
 
                 telemetry.ResultCode = requestTaskStatus.ToString();