+++ WEB/Src/DependencyCollector/Shared/Implementation/EventHandlers/DiagnosticsEventHandlerBase.cs
@@ -77,13 +77,7 @@
                 telemetry.Context.Operation.ParentId = activity.ParentId;
             }
 
-            foreach (var item in activity.Tags)
-            {
-                if (!telemetry.Properties.ContainsKey(item.Key))
-                {
-                    telemetry.Properties[item.Key] = item.Value;
-                }
-            }
+            this.PopulateTags(activity, telemetry);
 
             foreach (var item in activity.Baggage)
             {
@@ -93,7 +87,21 @@
                 }
             }
 
-            telemetry.Success = this.IsOperationSuccessful(eventName, eventPayload, activity);
+            if (!telemetry.Success.HasValue || telemetry.Success.Value)
+            {
+                telemetry.Success = this.IsOperationSuccessful(eventName, eventPayload, activity);
+            }
+        }
+
+        protected virtual void PopulateTags(Activity activity, OperationTelemetry telemetry)
+        {
+            foreach (var item in activity.Tags)
+            {
+                if (!telemetry.Properties.ContainsKey(item.Key))
+                {
+                    telemetry.Properties[item.Key] = item.Value;
+                }
+            }
         }
 
         protected virtual string GetOperationName(string eventName, object eventPayload, Activity activity)