+++ WEB/Src/DependencyCollector/Shared/Implementation/ProfilerSqlCommandProcessing.cs
@@ -2,6 +2,7 @@
 {
     using System.Data;
     using System.Data.SqlClient;
+    using Microsoft.ApplicationInsights.DataContracts;
     using Microsoft.ApplicationInsights.DependencyCollector.Implementation.Operation;
     using Microsoft.ApplicationInsights.Extensibility;
 
@@ -13,7 +14,7 @@
         /// <summary>
         /// Initializes a new instance of the <see cref="ProfilerSqlCommandProcessing"/> class.
         /// </summary>
-        internal ProfilerSqlCommandProcessing(TelemetryConfiguration configuration, string agentVersion, ObjectInstanceBasedOperationHolder telemetryTupleHolder)
+        internal ProfilerSqlCommandProcessing(TelemetryConfiguration configuration, string agentVersion, ObjectInstanceBasedOperationHolder<DependencyTelemetry> telemetryTupleHolder)
             : base(configuration, agentVersion, telemetryTupleHolder)
         {            
         }